Split Counter slice and Auth slice into separate files - making them more managable
