+++ src/components/Counter.js
@@ -1,7 +1,7 @@
 import { useDispatch, useSelector } from 'react-redux';
 import classes from './Counter.module.css';
 import React from 'react';
-import { counterActions } from '../store/index';
+import { counterActions } from '../store/counter';
 
 const Counter = () => {
